feat: start create tsdoc and optimize code
