+++ .eslintrc.json
@@ -15,7 +15,8 @@
     }
   },
   "plugins": [
-    "@typescript-eslint"
+    "@typescript-eslint",
+    "eslint-plugin-tsdoc"
   ],
   "settings": {
     "noInlineConfig": true,
@@ -37,6 +38,7 @@
     "linebreak-style": "off",
     "node/no-unsupported-features/es-syntax": "off",
     "no-underscore-dangle": "off",
-    "import/prefer-default-export": "off"
+    "import/prefer-default-export": "off",
+    "tsdoc/syntax": "warn"
   }
 }
